data_app/requirements.txt
pandas==1.2.3
streamlit==0.84.2
google-cloud-storage==1.41.0
